Flattening & translating
#3 Adding a translation of the json format corresponding to the locale file on the inverter

The selection of the locale could be made optional if needed (locales I could confirm working: en-US, de-DE, nl-NL, fr-FR

Chose to actually do the translation and not maintaining the key. As we are flattening the json file, some values are to be translated as well. We couldn't flatten it completely if we kept the keys as you wouldn't be able to distinguish values from translation id's